literalisan minden

+++ HotelGuru/HotelGuru.DataContext/Entities/Room.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Azure.Identity;
 
 namespace HotelGuru.DataContext.Entities
 {
@@ -10,7 +11,7 @@ namespace HotelGuru.DataContext.Entities
     {
         public int Id { get; set; }
         public string Type { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public bool Avaible { get; set; }
         public int UserId { get; set; }
         public Reservation Reservation { get; set; }
